fix: arm mask CV_64F→CV_32F type mismatch and add debug panel toggle
- Change createArmMask() Mat type from CV_64F to CV_32F to match the
  rest of the blend pipeline, fixing Core.multiply crash
- Add expand/collapse toggle button for debug log panel (180dp ↔ 60% screen)

+++ app/src/main/java/com/virtualfittingroom/warp/ClothingWarpEngine.kt
@@ -364,7 +364,7 @@ class ClothingWarpEngine {
      */
     private fun createArmMask(pose: PoseTracker.BodyPose, w: Int, h: Int, targetSize: Size): Mat? {
         return try {
-            val armMask = Mat.zeros(h, w, CvType.CV_64F)
+            val armMask = Mat.zeros(h, w, CvType.CV_32F)
 
             val armThickness = max(10, (pose.shoulderWidth * w * 0.1).toInt())
             val jointRadius = max(8, (pose.shoulderWidth * w * 0.06).toInt())
